pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loofi_fedora_tweaks"
version = "2.4.0"
description = "A comprehensive Fedora Linux system management tool with GUI, CLI, and daemon modes"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Loofi" }]
keywords = ["fedora", "linux", "system-management", "pyqt6", "tweaks"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "PyQt6>=6.7",
    "requests>=2.31",
    "dbus-python>=1.3",
]

[project.optional-dependencies]
api = ["fastapi>=0.128", "uvicorn>=0.30", "PyJWT>=2.8", "bcrypt>=4.1", "httpx>=0.27", "python-multipart>=0.0.9"]
dev = ["pytest>=8.0", "flake8>=7.0", "mypy>=1.10", "pytest-cov>=5.0", "pytest-timeout>=2.3"]

[project.urls]
Homepage = "https://github.com/loofitheboss/loofi-fedora-tweaks"
Repository = "https://github.com/loofitheboss/loofi-fedora-tweaks"
"Bug Tracker" = "https://github.com/loofitheboss/loofi-fedora-tweaks/issues"
Documentation = "https://github.com/loofitheboss/loofi-fedora-tweaks/tree/master/docs"

[project.scripts]
loofi-fedora-tweaks = "loofi_fedora_tweaks.main:main"

[tool.setuptools]
packages = ["loofi_fedora_tweaks"]

[tool.setuptools.package-dir]
loofi_fedora_tweaks = "loofi-fedora-tweaks"

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 60
addopts = "-v --tb=short"

[tool.mypy]
ignore_missing_imports = true
warn_return_any = true
no_error_summary = true

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["loofi_fedora_tweaks"]
